+++ src/InstrumentClip.cpp
@@ -2612,10 +2612,11 @@ expressionParam:
 						if (error) return error;
 
 						if (expressionParams) {
-							// Most other times you don't have to think about calling this. It's just because we didn't
-							// know which ParamCollection we were gonna load into, and MIDIParamCollection doesn't keep
-							// track of automation.
-							if (param->isAutomated()) expressionParams->paramHasAutomationNow(summary, paramId);
+							if (param->isAutomated())
+								expressionParams->paramHasAutomationNow(
+								    summary,
+								    paramId); // Most other times you don't have to think about calling this. It's just because we didn't know which
+							// ParamCollection we were gonna load into, and MIDIParamCollection doesn't keep track of automation.
 
 							// If channel pressure, gotta move and scale the values from how they were in the pre-V3.2 firmware
 							if (paramId) {
@@ -3311,14 +3312,14 @@ int InstrumentClip::claimOutput(ModelStackWithTimelineCounter* modelStack) {
 			bool paramManagerCloned = false;
 
 			// Maybe we (cryptically) marked it as "no drum".
-			if ((uint32_t)thisNoteRow->drum == (uint32_t)0xFFFFFFFF) {
+			if ((uintptr_t)thisNoteRow->drum == (uintptr_t)0xFFFFFFFF) {
 				thisNoteRow->drum = NULL;
 			}
 
 			// Or a gate drum from a pre-V2.0 Song file...
-			else if ((uint32_t)thisNoteRow->drum > (uint32_t)(0xFFFFFFFE - NUM_GATE_CHANNELS)) {
+			else if ((uintptr_t)thisNoteRow->drum > (uintptr_t)(0xFFFFFFFE - NUM_GATE_CHANNELS)) {
 
-				int gateChannel = 0xFFFFFFFE - (uint32_t)thisNoteRow->drum;
+				int gateChannel = 0xFFFFFFFE - (uintptr_t)thisNoteRow->drum;
 
 				thisNoteRow->drum = kit->getGateDrumForChannel(gateChannel);
 
@@ -3338,7 +3339,7 @@ int InstrumentClip::claimOutput(ModelStackWithTimelineCounter* modelStack) {
 			else {
 
 				// Don't call setDrum(), because that would overwrite the NoteRow's paramManager. It already has the right one, loaded from file
-				Drum* drumFromIndex = kit->getDrumFromIndex((uint32_t)thisNoteRow->drum);
+				Drum* drumFromIndex = kit->getDrumFromIndex((uintptr_t)thisNoteRow->drum);
 
 				ParamManagerForTimeline* otherParamManager;
